silent compilation

+++ Generator/INC/file.h
@@ -7,6 +7,6 @@
 #ifndef FILE_H
 #define FILE_H
 
-void createNDex(char* ip, int port, long long packSize, char* fileName);
+void createNDex(char* ip, int port, int packSize, char* fileName);
 
 #endif /* FILE_H included */
